+++ pkg/barback/test/transformer/bad.dart
@@ -22,7 +22,7 @@ class BadTransformer extends MockTransformer {
 
   BadTransformer(this.outputs);
 
-  Future<bool> doIsPrimary(Asset asset) => new Future.value(true);
+  Future<bool> doIsPrimary(AssetId id) => new Future.value(true);
 
   Future doApply(Transform transform) {
     return newFuture(() {