+++ pkg/barback/test/transformer/lazy_rewrite.dart
@@ -20,7 +20,7 @@ class LazyRewriteTransformer extends RewriteTransformer
   Future declareOutputs(DeclaringTransform transform) {
     if (consumePrimary) transform.consumePrimary();
     for (var extension in to.split(" ")) {
-      var id = transform.primaryInput.id.changeExtension(".$extension");
+      var id = transform.primaryId.changeExtension(".$extension");
       transform.declareOutput(id);
     }
   }