+++ sdk/lib/_internal/pub/lib/src/barback/excluding_transformer.dart
@@ -37,19 +37,19 @@ class ExcludingTransformer extends Transformer {
 
   ExcludingTransformer._(this._inner, this._includes, this._excludes);
 
-  Future<bool> isPrimary(Asset asset) {
+  Future<bool> isPrimary(AssetId id) {
     // TODO(rnystrom): Support globs in addition to paths. See #17093.
     if (_includes != null) {
       // If there are any includes, it must match one of them.
-      if (!_includes.contains(asset.id.path)) return new Future.value(false);
+      if (!_includes.contains(id.path)) return new Future.value(false);
     }
 
     // It must not be excluded.
-    if (_excludes != null && _excludes.contains(asset.id.path)) {
+    if (_excludes != null && _excludes.contains(id.path)) {
       return new Future.value(false);
     }
 
-    return _inner.isPrimary(asset);
+    return _inner.isPrimary(id);
   }
 
   Future apply(Transform transform) => _inner.apply(transform);